Remove utility 'isEnvironment' in favor of `instanceof`.
The 'isEnvironment' utility function was created so that webreed environments from
different versions of the 'webreed-core' package could be used in harmony.

By using `instanceof` assertion errors will occur when a plugin or user script attempts
to mix multiple versions of the `webreed-core` package within the same project context.
This should make it easier to diangose and thus update outdated plugins and/or user
scripts.